refactor(web): reorganize components for #560
