+++ webserver/web/src/components/app/information-banner.ts
