+++ webserver/web/src/components/app/stats-panel.ts
