+++ webserver/web/src/components/app/version-footer.ts
