+++ webserver/web/src/components/flags/sync-flags-button.ts
@@ -2,8 +2,8 @@ import { LitElement, html, css } from 'lit';
 import { customElement, state } from 'lit/decorators.js';
 import { createPromiseClient, PromiseClient, Interceptor, ConnectError } from '@connectrpc/connect';
 import { createConnectTransport } from '@connectrpc/connect-web';
-import { ConfigService } from '../gen/image_processing_connect';
-import { telemetryService } from '../services/telemetry-service';
+import { ConfigService } from '../../gen/image_processing_connect';
+import { telemetryService } from '../../services/telemetry-service';
 
 const tracingInterceptor: Interceptor = (next) => async (req) => {
   const headers = telemetryService.getTraceHeaders();
