+++ webserver/web/src/domain/interfaces/IProcessorCapabilitiesService.ts
@@ -1,5 +1,5 @@
 import type { FilterDefinition } from '../../gen/common_pb';
-import type { Filter } from '../../components/filter-panel.types';
+import type { Filter } from '../../components/app/filter-panel.types';
 
 export interface IProcessorCapabilitiesService {
   initialize(): Promise<void>;
